+++ apps/commerce-api/src/main/java/com/loopers/application/like/LikeEventPublisher.java
@@ -1,6 +1,5 @@
-package com.loopers.application.catalog.product;
+package com.loopers.application.like;
 
-//TODO 이거 수정해야함..
 public interface LikeEventPublisher {
   void increase(String userId, Long productId);
